consolidate darkmode logic in navbar component
